Update patch set 8
Patch Set 8:

I did want to use a generic mode parameter, but your argument changed my mind.  

You were right that using a generic mode would require the caller to know the modes.  Since the modes are implementation-specific, in this case, specific to GDB, that would complicate having code in the DSF plugins to request disassembly without opcodes, which is not very good.  Even if we don't need it right now, maybe later we'll want to be able to do that.

The alternative would be to define the modes in DSF instead of DSF-GDB and then map the DSF mode to the GDB mode in the service.  In our case the two sets of modes would be the same, but for someone not using GDB, they may actually need to do a mapping.

I wasn't sure this was really necessary considering the low likely hood of new modes being defined.

But you have more experience with the feature: what do you prefer?  If you like the generic mode idea, can you post an updated patch?

Patch-set: 8