Update patch set 1
Patch Set 1: Code-Review-1

The proposed change makes sense, i.e. adding the ability to specify a clip region when drawing text. However, this should be an optional argument, not a required argument, since making it required changes the API, and also complicates the simple case where one doesn't need a clipping region. (And yes, it might be a design choice to let a given label overlap a bit when the graphic object gets small, for clarity.)

A simple fix would be to add copies of the original versions of the GUIUtils drawText functions, without the clip argument, as overloads. That way you can keep the changes to the visualizer graphic object classes, since those look fine.

Also, don't forget to add yourself as a contributor in
the header comment!

Patch-set: 1
Label: Code-Review=-1