Update patch set 1
Patch Set 1:

Good point, but for example the component might also have visual substructure that's not represented as separate graphic elements. Providing the mouse location allows the
component to decide what to do (and it can always punt and
return a constant string if that turns out to be the right
thing to do).

Patch-set: 1