Bug 429196 Eliminate StackOverflow on adaptBinding.
A pretty simple and kludgy fix to the problem, but if we detect that
we are adapting a binding that we are just in the middle of adapting,
we bail and return null. Added Andrew's JUnit that reproduces the
problem in case someone wants to try a better solution.

Change-Id: Ib4a85c161be6aee073fee7ac0501464b70020fac
Reviewed-on: https://git.eclipse.org/r/24396
Reviewed-by: Doug Schaefer <dschaefer@qnx.com>
IP-Clean: Doug Schaefer <dschaefer@qnx.com>
Tested-by: Doug Schaefer <dschaefer@qnx.com>