Update patch set 1
Patch Set 1: Code-Review-1

I'll defer to Marc (D. or K.) to review the DSF-related changes.

The visualizer changes look fine overall. My only question is why we're overriding getGraphicObject() in MulticoreVisualizerCanvas -- it doesn't appear to be called from anywhere in the modified code. If it's not needed for this patch then we shouldn't include it.

Regarding the question in the getGraphicObject() override, it's been a while but I believe the reason we don't use the m_object array from the superclass is because the MulticoreVisualizer has multiple graphic object types (cpus, cores, threads, etc.) and it was more efficient to just track these in separate lists in the derived type rather than using the superclass list -- the latter is basically a convenience anyway, for canvases that only need a single list in which to store their content objects.

Patch-set: 1
Label: Code-Review=-1