Update patch set 3
Patch Set 3:

True, in my earlier example,
my gdb based system will give:

p sizeof (void *)
$6 = 2            // 2 character units (bytes)
p/x (char)-1
$7 = 0xffff  //hex format, 2 octets

It is true that there is some functionality not inline with the above
and does not consider the addressable size as you have pointed out.

I have followed some of the code and you can see e.g. HexRendering#getNumCharsPerByte()
which hard codes the number of characters per byte to 2  :-(

We could hopefully fix most of the cases by simply updating readmemorybytes to offer octets but based on the above example, 
this may not be as simple, so I guess we need to tackle one by one.

I can open a separate bug to ivestigate/discuss this further
so we make a consistent use of GDBMemory#readAddressSize().

Based on your observations I am ok with your current solution as per patchset 3,
we could update it (if needed) at a later point based on the results from the investigations/decision 

Thank you !

/Alvaro

Patch-set: 3