Bug 433222 - Turn on Others category in New Proj Wizard by default.
With this preference off, the new project wizard is a mess. We need
to have categories at the top level, especially as the number of
categories grow. This also fixes an issue with the Autotools which
tried to go into the Others category, but made a bit of a mess and
doesn't really need to be there anyway.

Change-Id: Idc67096756caa77087acabacb18ada41a43202b4
Reviewed-on: https://git.eclipse.org/r/25375
Tested-by: Hudson CI
Reviewed-by: Doug Schaefer <dschaefer@qnx.com>