Update patch set 1
Patch Set 1:

Hi Jeff. I'm excited to see this going into master. Is this change targeted for Luna?

I was thinking the doc part could come in a separate patch. I'm not sure it's a good idea to duplicate all this stuff, it's really big. Would it not be possible to include the regular CDT doc in addition to only the documentation specific to the RCP in an additional doc plugin?

Patch-set: 1
Reviewer: Gerrit User 598 <598@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0