Update patch set 1
Patch Set 1:

> Hi Jeff. I'm excited to see this going into master. Is this > change targeted for Luna?

Yes, I had permission to get this in for M7 from Doug and M7 is feature freeze.

> I was thinking the doc part could come in a separate patch. > I'm not sure it's a good idea to duplicate all this stuff, > it's really big. Would it not be possible to include the
> regular CDT doc in addition to only the documentation  
> specific to the RCP in an additional doc plugin?

Unfortunately, there were quite a number of changes required to the CDT docs as certain items have been removed and I needed to emphasize where things were done automatically on behalf of the end-user.  It's not the most ideal solution ever, but it was the best I could come up with.  The audience doesn't want to learn too much about Eclipse, CDT, etc... 

I am working on updating this patch as I hit the send to gerrit a bit too quick.  I am adding the script stuff so it can run from the M7 CPP EPP.  I hope to post it tomorrow.

Patch-set: 1
Label: Verified=0