Terminal: Bug 434487 - Key bindings for "maximize view" and "quick access"
This initial commit contributes "Alt+Up" for "Maximize View",
it also works for "Restore View"; as well as "Alt+Right" for
"Quick Access" in lieu of Ctrl+3 which may need to go to the Terminal.

Still TBD: Add those commands to the context menu such that they are
discoverable; and, decide whether "Alt+Down" should also be contributed
as a more intuitive variant for "restore view".

Signed-off-by: Martin Oberhuber <martin.oberhuber@windriver.com>
Change-Id: Iebf912cb404636c43014e15af9e63dea407e083f