Update patch set 2
Patch Set 2:

>  What guarantees that GDBHardareAndOS will get this 
> even[t] before SessionOSData, so that the latter 
> already sees isAvailable set to true?

Good question.  Note that SessionOSData does not check isAvailable() when getting DataModelInitializedEvent so technically no problem (and you will see below that this call would be unnecessary anyway).  However, it would seem bad for SessionOSData to make a call to the service if the service didn't know itself it was ready (one can imagine the service replying with an error in that case).

This is not a problem because DSF events are always sent to DsfServices first.  One can see this in DsfSession#doDispatchEvent() where the list of listener is sorted with DsfServices coming first.

Patch-set: 2