Update patch set 2
Patch Set 2: Verified+1 Code-Review+2

Looks like GDBHardwareAndOS#isAvailable() returning false all the time was an oversight when GDBHardwareAndOS was made to implement IGDBHardwareAndOS2 in bug 396268.  It went unnoticed because only SessionOSData was making use of IGDBHardwareAndOS2#isAvailable() and SessionOSData was only meant to work with GDB >= 7.5.

But IGDBHardwareAndOS2#isAvailable() should return the proper response for GDB < 7.5, which this patch does well.

Patch-set: 2
Label: Verified=+1
Label: Code-Review=+2