Bug 435141 - Split CDT native code fragments into a separate feature
This commit creates a new feature "org.eclipse.cdt.platform.fragments"
which is included by the cdt.platform feature and only holds the
CDT native code fragments.

Nothing changes for existing consumers of the cdt.platform feature or
cdt.sdk feature, since the new feature is included. But clients who
only want to consume a subset of CDT functionality can now pull in
that subset by import-package or require-bundle and add the required
fragments by depending on the new cdt.platform.fragments feature.

This allows clients to provide add-ons to CDT core without having to
pull in all of the CDT UI.


Change-Id: Ib8d1487ade923ba09544cfedd2c280b6ea1e10f0
Signed-off-by: Martin Oberhuber <martin.oberhuber@windriver.com>