Update patch set 1
Patch Set 1: Verified+1 Code-Review+1

Good to know this render is also available, 

just and observation, I find this render's name a bit confusing as it applies the endianess representation as per the number of octets in the column, e.g. not necessarily the size of an integer. However this is the best it can be done respecting the actual hex contents of the memory.

Patch-set: 1
Reviewer: Gerrit User 482 <482@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=+1
Label: Code-Review=+1