Bug 435886 - Update cdt.doc.isv
The org.eclipse.cdt.doc.isv bundle had not been updated since Juno.
This commit adatps to the changes from bug 303083, and does some more
updates to comply with CDT Luna:

1. Update version number - I deliberately jumped from 5.2.0 to 5.4.0
   in order to leave room for a Kepler version (not sure if that is
   a good idea to leave out a version)
2. Update CDT Release name (8.4.0) and Coppyright Date (2014)
3. Update link to Oracle Java Docs since BREE is now Java 7
4. Add org.eclipse.cdt.core.native to the sourcepath for Javadocs
5. Fix ProcessList extension point which now lives in o.e.cdt.native
6. Add spawner and pty packages to API docs

There might be more CDT API packages that our community currently
uses and that are not yet generated into the API docs. I only chose
to add spawner and pty since those happen to be the ones that I am
familiar with.

Change-Id: I6c16aed4408d83994075a9b0505bc7d3e68b11d1
Signed-off-by: Martin Oberhuber <martin.oberhuber@windriver.com>