Bug 435887 - Update cdt.doc.isv
The org.eclipse.cdt.doc.isv bundle had not been updated since Juno.
This commit adapts to the changes from bug 303083, and does some more
updates to comply with CDT 8.4 Luna:

1. Update the doc.isv bundle version number - I deliberately jumped
   from 5.2.0 to 5.4.0 in order to leave room for a Kepler version
   (not sure if it is a good idea to leave out a version)
2. Update CDT Release name to "8.4 (Luna)" and Coppyright Date (2014)
3. Update link to Oracle Java Docs since BREE is now Java 7
4. Add org.eclipse.cdt.core.native to the sourcepath for Javadocs
5. Add o.e.cdt.util.spawner and o.e.cdt.util.pty packages to API docs
6. Remove org.eclipse.cdt.core/search from the source path list
   That source folder had been removed 8 years ago when PDOM came :)
7. Fix an XHTML validation error in about.html that made it not
   possible to view Legal Notices from the docs link

There might be more CDT API packages that our community currently
uses and that are not yet generated into the API docs. Adding them
should be easy if that's desired, but I noticed that in many cases
the Javadocs themselves are somewhat weak so I didn't want to expose
them without the code owner's approval.

Change-Id: I6c16aed4408d83994075a9b0505bc7d3e68b11d1
Signed-off-by: Martin Oberhuber <martin.oberhuber@windriver.com>