Bug 428658 - fix keyboard-interactive authentication
Also removes unnecessary code for firstTry for password/
keyboard-interactive. And fixes firstTry for pubKey auth.
JSchConnection.newSession is storing the password with
session.setPassword. This password is used first before
jsch is using the UserInfo to prompt. Thus it isn't
necessary to return from the prompt first the same
stored password.

Change-Id: I17a4ae057fc595c2afe452c550bf47b70b696ef4