Update patch set 4
Patch Set 4:

Removing testAbstractBuiltinSpecsDetector_EnvChangesConfiguration_1 is probably OK (I take that as confirmation that this situation does not occur in practice), but I am not convinced that your changes to testAbstractBuiltinSpecsDetector_EnvChangesConfiguration_2 are correct. Maybe you can explain?

It seems to me like the test just misses the problem now. By no longer resetting isExecuted before every change, you no longer check whether the provider actually executed (which in fact it didn't), where by "executed" I mean not just a call to AbstractBuiltinSpecsDetector.execute(), but also an execution of the corresponding job that sets isExecuted to true and (among others) calls AbstractBuiltinSpecsDetector.runForEachLanguage(), where in the real thing the execution of GCC happens. The reason why you still get the correct value out of provider.getSampleEnvVar() is that setting that happens in MockBuiltinSpecsDetectorEnvironmentChangeListener.execute(). If it happened in MockBuiltinSpecsDetectorEnvironmentChangeListener.runForEachLanguage(), which would be a more realistic setup, the test would fail.

It seems to me that the underlying problem, which we still haven't solved, is that a MockBuiltinSpecsDetectorEnvironmentChangeListener, or any other AbstractBuiltinSpecsDetector that is not an ILanguageSettingsEditableProvider, can belong to multiple configurations at the same time (because it is not cloned in LanguageSettingsProvidersSerializer.cloneProviders()), but still has only one currentCfgDescription[Cache] that we are using to judge its currentness. Am I understanding this correctly?

Patch-set: 4
Reviewer: Gerrit User 858 <858@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0