Update patch set 3
Patch Set 3:

Hi Bill,

In your comment about interface cleanliness regarding the dispose() method in IVisualizerModelDataSource: Would you elaborate a bit about what you meant? 

Is it the idea of having a cleanup method in the interface or the fact that it's called "dispose"? 

Another thought: In the case of the DSFSessionState implementation of the interface, calling dispose() is probably not necessary, as long as RemoveAllServiceEventListeners() is called. is that perhaps what you meant?

Thanks,

Marc

Patch-set: 3