Update patch set 3
Patch Set 3:

> In your comment about interface cleanliness regarding the > dispose() method in IVisualizerModelDataSource: Would you > elaborate a bit about what you meant?

> Is it the idea of having a cleanup method in the interface...

Right, ideally an interface should be only the behavior that's essential to the interface. Now, there's nothing saying an interface can't have a cleanup method, e.g. "close()" on an IO interface, or "release()" on an object
that holds some resource, but "dispose()" here didn't seem essential as part of this particular interface definition.

(BTW, I have no problem with "dispose()" as the name of
the method, since I'm where the usage came from!)

> Another thought: In the case of the DSFSessionState
> implementation of the interface, calling dispose() is
> probably not necessary, as long as
> RemoveAllServiceEventListeners() is called. is that
> perhaps what you meant?

Welll... it's more like if you have a list of the objects
in question via a "real" reference through which dispose()
can be called, then dispose() doesn't need to be on the
interface. If you're using the interface so that you can have a bunch of objects of various types held only by a reference to the interface, then you might need to expose dispose() as well on the interface to allow them to be disposed generically. (Though in that case, technically the
dispose() method would want to be on an IDisposable interface that IVisualizerModelDataSource extends. Or would that be _too_ C#? ;-)

Patch-set: 3