Update patch set 4
Patch Set 4: Code-Review-1

Hi Iulia, 
It is preferred to display an warning that file will be changed and actually redo the project link file to refer to newly imported executable.
Most likely the use-case is for an user to debug again a different version of application that is from a different folder. 
Allowing user knowingly to replace the old link is better then forcing user to exit wizard, delete old link from the project and start the wizard again.

Patch-set: 4
Label: Code-Review=-1