Bug 437392 Initial commit of LaunchBar.
The LaunchBar is a Tool Control that allows the user to select an
active launch configuration, launch mode, and a new concept, a launch
target, and the build for that combination and launch that combination.
It also provides shortcuts to create launch configurations, and edit
their properties using the existing launch config edit dialog.

Change-Id: I8c6c9b72fed9052c03d2028c4611c78872d971c4