Update patch set 3
Patch Set 3:

Hi Bill,

Thanks for the clarifications. 

"If you're using the interface so that you can have a bunch of objects of various types held only by a reference to the interface, then you might need to expose dispose() as well on the interface to allow them to be disposed generically." 

Yes, that's my case here. I want to be able to call dispose() using IVisualizerModelDataSource references. See MulticoreVisualizer#disposeDataSource(). 

"(Though in that case, technically the dispose() method would want to be on an IDisposable interface that IVisualizerModelDataSource extends. Or would that be _too_ C#? ;-) "

This seems like a good way to go. I'll push a new patchset that included this. 

One question: it happens that there is an existing IDisposable interface (defined in package org.eclipse.ui.services), but it was meant for something not quite related ("services that make themselves available through the IAdaptable mechanism"). 

Technically that interface has what I need (i.e. dispose() method), but my use of it seems a bit outside of what was intended. Do you think I should instead create my own IDisposable? 

Thanks again!

Marc

Patch-set: 3
Label: Verified=0