Bug 437525: NPE in QMakeProjectInfoManager
This fixes the NPE but not the underlying reason for the Qt plugin to be
activated.  I think that might be happening because of the Codan checker
-- from what I recall, that extension point doesn't have a way to
prevent plugin activation.

The NPE is fixed by checking CACHE for null before using it.  The #stop
method unregisters and sets CACHE to null.  However, unregistering
doesn't prevent notifications that are already in flight.  I suspect
that notification arrives, and then CACHE is null.

My test workspace would throw the NPE on every exit, with this change it
never does.

Change-Id: I01360a4c19c85fba53269d265948daf6d083c8d1
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>