Update patch set 2
Patch Set 2:

(12 comments)

Thanks Vladimir, your input is very welcomed. I have responses in line, but we should really move some of this discussion to the bug since I've already submitted this request and wasn't really looking at it any more.

I've been pushing directly until I get it to a point where it's ready to be extended. I'll push to gerrit after that. If you guys have anything you'd like to see and work on, I'd be happy to look at requests as well.

Patch-set: 2