Update patch set 1
Patch Set 1:

So are we able to run the display.asyncExec with this new Display?  Does it matter that we do or don't if Eclipse is exiting?

With your proposed change, we wouldn't be getting this new display anymore so technically the behaviour will change.

Do we know why this issue happens with the DsfSourceDisplayAdapter and not with other areas that use Display.getDefault()?  It is used all over.

Patch-set: 1