Update patch set 1
Patch Set 1: Code-Review+2

I'm not concerned about the other places.  I'm concerned that if the problem is only seen here, then there is something else going on and the solution may not be correct.

However, the change does look safe so I agree that we might as well put it in and leave it at that.

Patch-set: 1
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+2