Update patch set 2
Patch Set 2:

"So are we able to run the display.asyncExec with this new Display? Does it matter that we do or don't if Eclipse is exiting?"

both methods (getDefault and getWorkbench().getDisplay()) do execute the asyncExec correctly when not exiting (removing debug sessions) and both don't execute when exiting Eclipse and a session was not removed. I haven't noticed any side effect after restarting Eclipse (annotations or current line background painter).

"With your proposed change, we wouldn't be getting this new display anymore so technically the behaviour will change."

The new display created by getDefault actually was a leak since it's not getting disposed. But in practice, Windows, Linux and OS X implementations of SWT don't allocate resources for the display. But SWT is free to change that and it could be the case on a different platform.

"Do we know why this issue happens with the DsfSourceDisplayAdapter and not with other areas that use Display.getDefault()? It is used all over."

It's because it's being executed during the bundle stop which happens after the default display getting nulled out. It could actually happen in other places but there are so many Display.getDefault. I think a reasonable approach is to keep this potential problem in mind and fix them as we go when we notice a potential issue. I also found out about Display.disposeExec that we could use when it's absolutely required that something gets executed on the UI thread before exit.

Patch-set: 2