Update patch set 4
Patch Set 4: Code-Review-1

I ran into one problem when dealing with multiple sessions.

1- start two debug sessions
2- change the format for one register
3- select the second debug session
=> the register for which the format was changed still shows the old format, although if I go to the context-menu to change that format, the new one is selected.

The problem is that #2 will cause a refresh of the register view for the current session, not for all sessions.  So other sessions won't know something has changed.  This bug was part of the PDA example it seems.

I have to think about how to handle this.

Patch-set: 4
Label: Code-Review=-1