Update patch set 5
Patch Set 5:

"would you know for which cases would this list of targets needs to be updated ?"

If you build the target as a package (a plain pom module here), the target file won't get included in the package if you only specify it on the command line. For example, if you go to the releng/o.e.cdt.target and do a mvn install, the target module will be packaged and installed to your local reactor in ~/.m2/repository/org/eclipse/cdt/... But only the files explicitly attached in the pom.xml will be there. Then you can use that built target in the reactor to build any submodules (well, provided that you also build the dependencies).

I actually wasn't expecting it to work without adding the target to the pom.xml but it's just a nice side-effect because then you can have your own custom targets and not have to modify the pom.xml, provided that you don't need it to be part of the target module package. If we eventually deploy the target module to Nexus then it becomes even more important that all the targets we want to be released are in the pom.xml.

Patch-set: 5
Reviewer: Gerrit User 598 <598@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0