Update patch set 4
Patch Set 4:

"Isn't undo support will happened automatically because we using re-writer here?"

It undoes but it goes back to the original state of the editor before the edit, not to what it would have been if the format action was disabled. This is what's missing from the JDT implementation. But perhaps it would make sense to do that in a future patch instead.

Patch-set: 4
Reviewer: Gerrit User 598 <598@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0