Bug 440710 - [ASTRewriter] UnhandledASTModificationException when ...
When a statement has been inserted into a replaced compound statement
createModifiedChildArray is called for the new compound statement. The
insert modification is attached to its sibling node. The implementation
for handling this insert was on the parent (compound statement) though.
I moved this (adapted) functionality to handling of the sibling nodes.
With the way insert modifications are stored the previous implementation
cannot have worked. Furthermore, there is no test code which covers the
deleted code.

Change-Id: I9615787177c12a870df7c9663170f38b2c28accb
Signed-off-by: Thomas Corbat <tcorbat@hsr.ch>