Update patch set 12
Patch Set 12:

Thanks Alvaro, great stuff!

Some in-line comments about the new service API.

Patchset 12 are minor changes I suggest we make to the current proposed solution.  You can compare patchset 12 with patchset 11 to see the changes.  The most noteworthy one is the change in GDBRegisters to allow IRegisters2.editRegisterGroup() to pass null to keep either the name or the list of registers the same.

Patch-set: 12
Label: Verified=0