Bug 437392: Major API changes to turn the Launch Bar Model a beauty
Pretty major changes but heavily reduce the size of the API. Check
out the ILaunchBarManager which now only supports providers. All other
methods are now for internal use by the UI plugin. Also change to the
model. The extension point has much more documentation now. Please see
it for an explanation. But essentially made launch configuration types
part of the model to make a nice cleaner bridge to the platform.

Change-Id: Icdd300e99f831c84cd0a25a23410aa1cd2ccfd7c