Bug 437392: Major API changes to turn the Launch Bar Model a beauty
Pretty major changes but heavily reduce the size of the API. Check
out the ILaunchBarManager which now only supports providers. All other
methods are now for internal use by the UI plugin. Also change to the
model. The extension point has much more documentation now. Please see
it for an explanation. But essentially made launch configuration types
part of the model to make a nice cleaner bridge to the platform.

Also includes change to Launch Targets view to make it a common
navigator so different target types can add objects to the tree.

Also a number of changes to try and delay plug-in loading with the
providers.

Change-Id: Icdd300e99f831c84cd0a25a23410aa1cd2ccfd7c
Reviewed-on: https://git.eclipse.org/r/33061
Tested-by: Hudson CI
Reviewed-by: Doug Schaefer <dschaefer@qnx.com>