Bug 437392: Target View using Common Nav, ILaunchTarget adaptable.
Switched the Launch Targets view to use the common navigator. This
will allow adoptors to add nodes to the tree specific to their target
types. For Arduino, may add one for the Serial Monitor to support
the Open action on it.

To make it easier to supply UI for the target, making ILaunchTarget
extend IAdaptable.

Change-Id: I55a3c8053415ddce76b31284f30f22609953eeae