Update patch set 5
Patch Set 5: Code-Review-2

Removing tests and writing new ones is not the same
as fixing existing ones,
you don't know now if tests I wrote before are passing.
You cannot say it is not applicable anymore because
most of them are still applicable with little tweaks.

I disagree that launch config should be saved unless
user edits it, because it pollutes workspace with configs
that nobody cares about

I disagree that all abstract classes are removed

I disagree with removal of launchObjectChanged, it allows to change descriptors, for example when project nature changed

Patch-set: 5