Update patch set 1
Patch Set 1:

It is because your commit does not have the changeId in it.

What you should do is take the changeId from the top left of the Gerrit page and amend your commit to include it.  Push to gerrit again and it will update the existing review.

Then abandon the new review.

Details here: 
https://wiki.eclipse.org/CDT/git#Pushed_without_a_changeId.3F

Patch-set: 1
Label: Code-Review=0