Bug 444196 - Add delay before deleting breakpoint
We are getting intermittend failures on the GDBConsoleBreakpointsTest
due to the fact that we create and then try to delete a breakpoint so
fast that the MIBreakpointsManager does not have enough time to update
its data structures.

This is something specific to automated tests, as a user won't be able
to delete a breakpoint so fast.

Change-Id: I4b220cc859b7ee45e1b8dbde0edbcaa38b9dfe34
Signed-off-by: Marc Khouzam <marc.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/33432
Tested-by: Hudson CI
Reviewed-by: Alvaro Sanchez-Leon <alvsan09@gmail.com>