Update patch set 1
Patch Set 1: Code-Review-1

Please provide an explanation of how you think this will solve the problem in all cases. It looks to me like you are just preventing conversion of path arguments to the remote path equivalent. Since the command is being executed on the remote system, any paths that are passed to it must also be correct.

There is a separate issue that RemoteCommandLauncher#makeRemote() assumes that a path is being passed to it. It should probably test Path.EMPTY.isValidPath(path) before doing the conversion.

Patch-set: 1
Label: Code-Review=-1