Update patch set 3
Patch Set 3:

Please explain the logic in documentation.  The RemoteCommandLauncher is converting the working path to remote, but it does not do so with the command path.  That should be fine for all relative commands, but in this case, the command is an absolute path from the local system.  The old logic in Autotools was checking for an absolute path being passed and then trying to convert it, assuming that running such an absolute command on the remote system might result in failure since the path might not exist.  Does the RemoteCommandLauncher know it is running a local absolute path and converts it automatically?  If so, this should be documented somewhere.

Patch-set: 3