Update patch set 3
Patch Set 3:

When building the autotools project it will get in RemoteCommandLauncher#constructCommandArray that will construct the command array to be passed to the process. Method constructCommandArray will call the method makeRemote that expects as parameter an IPath localPath. As we used the method getRemotePath before, we already converted the local path to a remote path and when the method makeRemote is called we will pass the remote path for it, not the local one. 

When the method makeRemote receive the remote Path instead of the local path, the variable: IPath relativePath = localPath.makeRelativeTo(remoteLocation); got a weird path to the autotools configure file in the remote machine and cant find it

Patch-set: 3
Reviewer: Gerrit User 7248 <7248@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0