Update patch set 3
Patch Set 3:

To answer Jeff's question: yes the RemoteCommandLauncher#execute() method attempts to normalize all the strings passed in the args array to the remote versions. The runScript() method was converting the first of these strings to the remote version before calling execute(), so the conversion was happening twice. For some reason, this worked on some platforms but not others. The Javadoc for execute() should be updated to mention the conversion.

There is a separate but related issue that the conversion assumes that all the arguments are paths. The makeRemote() method should check that the argument is actually a path before trying to convert it. This could be a separate patch though.

Patch-set: 3