Bug 444395 - Must synchronize creation of bp extension
It is possible that more than one thread try to fetch the extension of
the same breakpoint.  If that extension is not created yet, both threads
could end up creating two different ones by mistake.

We must synchronize the possible creation of the breakpoint extension to
avoid such a situation.

Change-Id: I51118bdfb9c4215fca14d7b5e18a5963e02cde34
Signed-off-by: Marc Khouzam <marc.khouzam@ericsson.com>