Update patch set 1
Patch Set 1: Code-Review-2

The IP policy still says you have to put your name at the top of the file in the copyright section.

Besides that the change looks good.

To submit a new patch, make the change, amend your commit and in the commit message, add the ChangeId that was created for this review, please see details here:
https://wiki.eclipse.org/CDT/git#Pushed_without_a_changeId.3F

Patch-set: 1
Label: Code-Review=-2