Bug 437392 - first cut at handling descriptor takeovers.
If a launchObject changes such that a new higher priority descriptor
type wants to take it over, we need to let it. This is handles on
the launchObjectChanged call.

I also added a note to the Project object provider as projects change
a lot and we only care about a subset of those events. Not sure how
to formalize that, though. Need to dig deeper into the deltas and
see what we can ignore.

Change-Id: I964d8785a474c43c917bcd7d8a84abf2b7904aa4