Update patch set 5
Patch Set 5:

> otherwise it may behave a bit worse than before this fix

The two issues are similar but I don't quite see why starting with this one would make things worse.

I think this fix reduces the number of possible race conditions, even if it does not fix them all.

Is there a scenario you have in mind that would be worse with this fix?

Patch-set: 5
Label: Verified=0