Update patch set 1
Patch Set 1:

"I wonder if we should enable the update site by default? People installing the CDT probably don't do it just for fun but actually want the plugins, so they would be interested in the CDT update site, no?"

I'm leaning towards not enabling it by default. I think it would only make sense to enable it by default if it is planned to have a 8.6.1 that is not aligned with the release train. People who want to install optional features (I assume that's what you meant by plugins) can enable it easily and then select the optional features. This is still more handy than having to look for the update site URL on the web. There is also a bit of overhead when having many updates sites enabled since I believe p2 will fetch the index of all enabled update sites when updating as well as when installing new features combined with "Contact all update sites".

Patch-set: 1