Update patch set 10
Patch Set 10:

Patchset 10 improves the handling of the ElementFormatEvent so that it can be prepared for more advanced VMProviders which set the format differently.  For example, setting the format of an array could set the format for all children of that array.  The new patch creates the delta properly based on the depth parameter of the event.

Patch-set: 10