Support proxy connection for gateway without sftp or with MaxSession=1
Command proxy does not work if both the host and the gateway have MaxSession=1
(caused by 356657).

Add support in JschConnection for a connection which has a session but isn't
fully setup. Such a connection isn't considered open. And if it is opened after
it already has a session without setupFully=false (default), then all setup
setups are done.

Also change that getProxyConnection returns null not a LocalConnection for
a local command. This reduces the dependency on LocalConnection from within
jsch.core.*.

Change-Id: I64d05e69749d96248d2f40bd2db021d809cb04a9