Bug 446104 - Fix target removal. Also fix Config Edit page.
The launch bar manager was missing the line to actually remove the
launch target in launchTargetRemoved. Added a test to detect that.

Also fixed the Config Edit page. Name verification was failing since
it was comparing against the false configs created when the tabs
are initialized. We clean up now right after that init.

Change-Id: I9cff816040f2e1866c9454cb00f69ec4468225d4
Reviewed-on: https://git.eclipse.org/r/34477
Tested-by: Hudson CI
Reviewed-by: Doug Schaefer <dschaefer@qnx.com>