[RFC] tests: breakpoint definition that does not hardcode line numbers
This is similar to what is done in gdb. A source line is annotated with a
comment, such as:

  printf ("Copy complete.\n");  /* breakpoint 2 */

then, the test look up the line number at runtime:

  set break2_loc [gdb_get_line_number "breakpoint 2"]

The big advantage is that you can add or remove lines to test programs without
fear of breaking everything.

This is more or less a proof of concept. If the idea is well-received, it
would probably be good to put the feature in some base class so that it is
available to more tests.

What do you think?

Change-Id: Id87e75487e46310f41736e63fd6da36df023c753
Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>