Update patch set 7
Patch Set 7: Code-Review-1

(4 comments)

I found a couple of issues while testing (actually, only one that I think we should fix).

I also noticed that when doing a detach from a local session, the process still gets killed eventually.  This is a problem even on master.  I believe it is because we have a console for the process in that case, and that means we've included the process in the platform's launch, so the platform probably does a destroy() on it.  I wanted your opinion on that and see if it is a problem for what you are trying to achieve.  After your input I can open a bug and we can deal with it separately.

Patch-set: 7
Label: Code-Review=-1