Make execution of tests.dsf.gdb configurable
When running maven builds it's sometimes desirable to validate the
compilation of the tests.dsf.gdb plugin but don't actually execute this
specific test cases but other ones which are part of the build reactor.
This small change introduces a new parameter (don't skip by default)
that allows skipping of this test cases if the parameter is specified in
the mvn command e.g. -Dskip.tests.dsf.gdb=true

Change-Id: I75dd28ee607ce19a7f03f03296fab3c8dbe7b964