Update patch set 1
Patch Set 1:

When running maven builds it's sometimes desirable to validate the compilation of the tests.dsf.gdb plugin but don't actually execute this specific test cases but other ones which are part of the build reactor.

This small change introduces a new parameter (don't skip by default) that allows skipping of this test cases if specified in the mvn command.

Patch-set: 1
Label: Verified=0