dsf-gdb tests: Improve tests Makefile
This changes the Makefile tha builds test apps so that it uses the
proper Makefile structures, rather than a single shell command.

Also, this will compile .c files with gcc and .cc files with g++,
allowing to have both.

Also, I dropped the .exe extensions. The justification "so that CVS does
not include it when making a patch" doesn't seem valid to me anymore.

Change-Id: I4414b1dc5c31a9eaa7edaed30e53363b9a76dd8f
Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>